Encapsulte Container Runtime DataStore Handling
